feat(chord): move stories to the dedicated storybook workspace
