+++ storybook/package.json
@@ -19,6 +19,7 @@
     "@nivo/bullet": "workspace:*",
     "@nivo/bump": "workspace:*",
     "@nivo/calendar": "workspace:*",
+    "@nivo/chord": "workspace:*",
     "@nivo/generators": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@babel/preset-env": "^7.21.5",
